Add CMakeFlag to disable OSCC faults
Prior to this commit, OSCC had a lot of its own propreitary faults. To assist with system diagnostics and characterisation, this commit provides a CMakeFlag which allows disabling all OSCC fault reports and handling.
It also expands the steering override flag to disable operator override on all firmware modules to remove more noise unnecessary to this process.